refactor(projects): move session logic to caller components
- Extract useSession from PostDetailCard to individual components
- Add isOwner prop to PostDetailCard for better reusability
- Update NoteDetail, OutcomeDetail, and ReviewDetail components
- Improve separation of concerns for authentication logic
