+++ entities/post/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './post-card';
+export * from './post-detail-card';
+export * from './post-preview-card';
