+++ features/projects/components/ReviewForm.tsx
@@ -1,7 +1,7 @@
 import { format } from 'date-fns';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import PostCard from '@/shared/components/Post/PostCard';
+import { PostCard } from '@/entities/post';
 import {
   projectReviewFormSchema,
   ProjectReviewFormSchema,
